--- src/utils/__init__.py
@@ -1,5 +1,0 @@
-from . import constants
-
-constants.IS_TEST = True
-
-from . import analysis, interfaces, routines, titration  # noqa: E402, F401
