+++ test/constants_test.py
@@ -2,22 +2,24 @@
 This test file is primarily to test the ability to import from the
 src module
 """
+# from titration import utils
+from titration.utils import interfaces, constants
 
-from .context import titration_module
-
+def setup_test_mode():
+    constants.IS_TEST = True
 
 def test_can_we_import():
-    assert titration_module.constants.IS_TEST is True
+    assert constants.IS_TEST is True
 
 
 def test_import_interfaces(capsys):
-    titration_module.interfaces.setup_interfaces()
+    interfaces.ui_lcd = interfaces.setup_lcd()
 
     # flush stdout
     _ = capsys.readouterr()
 
-    titration_module.interfaces.ui_lcd.print(
-        "We did it", 1, titration_module.constants.LCD_LEFT_JUST
+    interfaces.ui_lcd.print(
+        "We did it", 1, constants.LCD_LEFT_JUST
     )
 
     captured = capsys.readouterr()
