--- test/context.py
@@ -1,9 +1,0 @@
-"""
-Import context for tests
-"""
-import os
-import sys
-
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
-
-import src.utils as titration_module  # noqa: E402, F401
