+++ titration/utils/__init__.py
@@ -1,0 +1,5 @@
+from . import constants
+
+constants.IS_TEST = True
+
+from . import analysis, interfaces, routines, titration  # noqa: E402,
