+++ titration/utils/constants.py
@@ -107,7 +107,7 @@ calibration_data_format = {
     "vol_pump": 0,
 }
 
-CALIBRATION_FILENAME = "../data/calibration.json"
+CALIBRATION_FILENAME = "data/calibration.json"
 # temperature calibration values
 TEMPERATURE_REF_RESISTANCE = 4300.0
 TEMPERATURE_NOMINAL_RESISTANCE = 1000.0
